==> 1_tokenize.ipynb <==
import os
import json
import numpy as np
import pandas as pd
import torch
from tqdm.notebook import tqdm 

from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer, PorterStemmer
from indicnlp.tokenize import indic_tokenize
from nltk.corpus import stopwords

from deep_translator import GoogleTranslator

def split_long_sentence(sentence, max_length=3000):
    parts = []
    for i in range(0, len(sentence), max_length):
        part = sentence[i:i + max_length]
        parts.append(part)

    return parts

def convertToEnglish(sentences):
    eng = []
    for sentence_array in sentences:
        sentence = ''
        engText=''
        for i in range(len(sentence_array)):
            sentence = sentence + sentence_array[i]
        if len(sentence)>2000:
            small_sentences = split_long_sentence(sentence, max_length=2000)
            engPart = ''
            for i in range(len(small_sentences)):
                current_translate = GoogleTranslator(source='hi', target='en').translate(small_sentences[i])
                engPart = engPart + current_translate
        else:
            engPart = GoogleTranslator(source='hi', target='en').translate(sentence)
        engText=engText +engPart
        eng.append(engText)
    return eng

stop=open('stopwords/hindi.txt')
hindi_stop_words=[]
for x in stop:
    hindi_stop_words.append(x.strip("\n"))

def tokenize_hindi(sentence):
    tokens = indic_tokenize.trivial_tokenize(sentence, lang='hi')
    
    filtered_tokens = [token for token in tokens if token not in hindi_stop_words]

    return filtered_tokens

english_stop_words = set(stopwords.words('english'))
def tokenize_english(sentence):
    tokens = word_tokenize(sentence)
    
    filtered_tokens = [token for token in tokens if token not in english_stop_words]
    
    lemmatizer = WordNetLemmatizer()
    lemmatized_tokens = [lemmatizer.lemmatize(token) for token in filtered_tokens]
    
    stemmer = PorterStemmer()
    stemmed_tokens = [stemmer.stem(token) for token in lemmatized_tokens]

    stemmed_tokens = [token for token in stemmed_tokens if token.isalpha()]
    
    return stemmed_tokens

def process(file,sampleCount):
    print(f"Current file:{file}")
    with open(file) as f:
        data = json.load(f)
    df = pd.DataFrame(data)

    df=df.transpose()
    df = df.head(sampleCount)

    facts_and_arguments = []
    for i, row in df.iterrows():
        facts_and_arguments.append(df.loc[i]['segments']['facts-and-arguments'])
    df['facts-and-arguments'] = facts_and_arguments
    
    

    engSentencesCollection = []
    sentencesCollection =[]
    tokensCollection = []
    engTokensCollection = []
    judgeOpinionCollection = []
    for i, row in tqdm(df.iterrows(),total=len(df)):
        judge_sentences=[]
        sentences=[]

        paras = df.loc[i]['judge-opinion']
        for para in paras:
            sent = para.split('।')
            sent = [i for i in sent if len(i)!=0 and i!=' ']
            judge_sentences.extend(sent)
        
        paras = df.loc[i]['facts-and-arguments']
        for para in paras:
            sent = para.split('।')
            sent = [i for i in sent if len(i)!=0 and i!=' ']
            sentences.extend(sent)
        sentences = [i.strip() for i in sentences]

        english_sentences = convertToEnglish(sentences)
        tokens = [tokenize_hindi(sentence ) for sentence in sentences]
        english_tokens = [tokenize_english(sentence ) for sentence in english_sentences]
        
        judgeOpinionCollection.append(judge_sentences)
        tokensCollection.append(tokens)
        sentencesCollection.append(sentences)
        engTokensCollection.append(english_tokens)
        engSentencesCollection.append(english_sentences)

    df.drop(columns=['segments','district','case_number'],inplace=True)
    df['sentences'] = sentencesCollection
    df['tokens'] = tokensCollection
    df['english_tokens'] = engTokensCollection
    df['sentences_english'] = engSentencesCollection
    df['judge-opinion'] = judgeOpinionCollection
    return df

base_path = "./data/raw/"
write_path = "./data/summary/"
files = os.listdir(base_path)
try:
    os.mkdir(write_path)
except:
    pass
files = [f for f in files if ".json" in f]

dfs=[]

sampleCount = [1,1,1]
for file,count in zip(files,sampleCount):
    dfs.append(process(base_path+file, count))

df = pd.concat(dfs, ignore_index=True)
df.to_csv(f"{write_path}/data.csv",index=False)
