==> 3_GroupByCrime.ipynb <==
import pandas as pd

input_path="./data/prior_case/"
output_path="./data/prior_case/groups/"
df = pd.read_csv(f'{input_path}data_crime.csv')

grouped = df.groupby('CrimeType')

class_dict = {}
keywords=['theft','rape','drug','murder','assault']
for name, group in grouped:
    group.to_csv(f'{output_path}data_{keywords[name]}.csv',index=False)
