==> 6_dialect_mapping.ipynb <==
import os
import json
import numpy as np
import pandas as pd
import torch
from tqdm.notebook import tqdm 

def map_words(sentences, word_mapping):
    mapped_sentences = []

    for sentence in sentences:
        words = sentence.split()
        mapped_words = [word_mapping.get(word, word) for word in words]
        mapped_sentence = ' '.join(mapped_words)
        mapped_sentences.append(mapped_sentence)
    return mapped_sentences

def process(df,dict):
    df['sentences']=df['sentences'].apply(lambda x: map_words(x, dict))
    return df


base_path = "./data/summary/"
write_path = "./data/summary/"
files = os.listdir(base_path)

df=pd.read_csv(f"{base_path}/data.csv")
df['sentences']=df['sentences'].apply(eval)
dict=json.load(open(f"./models/dialect_dict.json","r"))
print(dict)
final_df=process(df,dict)
final_df.to_csv(f"{write_path}/data_ranked.csv",index=False)
