==> 8_summary_score.ipynb <==
import pandas as pd
import tqdm.notebook as tq
input_path ="./data/summary_results/"
from indicnlp.tokenize import indic_tokenize

def calculate_responsiveness(source_document, generated_summary):
    source_tokens = set(indic_tokenize.trivial_tokenize(source_document))
    summary_tokens = set(indic_tokenize.trivial_tokenize(generated_summary))
    overlapping_unigrams = source_tokens.intersection(summary_tokens)
    
    if len(source_tokens) > 0:
        responsiveness_score = len(overlapping_unigrams) / len(source_tokens) * 100
    else:
        responsiveness_score = 0
    
    return responsiveness_score

def get_all_files(input_path):
  import os
  files = os.listdir(input_path)
  return files
files = get_all_files(input_path)

for file in files:
  data = pd.read_csv(f'{input_path}{file}')
  data['sentences'] = data['sentences'].apply(eval)
  data['summary'] = data['summary'].apply(eval)
  data['summary_english'] = data['summary_english'].apply(eval)
  data['sentences']=data['sentences'].map(lambda x: ' '.join(x))
  data['summary'] = data['summary'].map(lambda x: ' '.join(x))
  data['summary_english'] = data['summary_english'].map(lambda x: ' '.join(x))
  scores = []
  scores_english = []
  for i in range(len(data)):
    scores.append(calculate_responsiveness(data['sentences'][i],data['summary'][i]))
    scores_english.append(calculate_responsiveness(data['sentences'][i],data['summary_english'][i]))
  score_value = sum(scores)/len(scores)
  score_english_value = sum(scores_english)/len(scores_english)
  print(f'{file}\nHindi {score_value}%\t English {score_english_value}%\t',end='\n')
